[boards] dmo: fix eth phy reset on dual lite
The RGMII pins on dual lite use different registers in comparison to quad/dual
core. This patch fixes two problems. First the phy is properly reseted and
second the m25p80 flash is detected. Using the wrong registers for the
phy reset caused wrong pin configuration for the spi flash.

Signed-off-by: Zahari Doychev <zahari.doychev@linux.com>