swu:make independent update-states from hardware
Hook function concept is used to make SWU more generic.
A swu_hook structure is defined in bbu.h in order to collect the HW-function
as a function pointer and the update state.

Two necessary functions as registering-hook and get-hook are defined in
swc.c. It means that the defeined swu-structure and update-state function
in HW(board.c) file will be registered and accessed via these two functions.

Now swupd.c can be independent from the HW. Thus, all HW-dependent functions
are replaced with the generic function, which is called by changing the update
states.

Those HW-dependent functions are moved into the board.c and the function is
registered during initilization process of the board.

Signed-off-by: Reyhaneh Yazdani <reyhane.y84@gmail.com>