swu: copy a dts as oftree if there is not
oftree means user customized-device-tree in our DMO.
When user needs to update display timing via SWU, it is required to change the
oftree file. By default there is no "oftree" in boot partition and we use our
standard device tress for booting.

our standard dts for solo/duallight:
- imx6dl-dmo-ppc-auo-g173hw01.dts
- imx6dl-dmo-ppc-auo-g804sn05.dts
- imx6dl-dmo-ppc-chimei-g070y2_12018614.dts
- imx6dl-dmo-ppc-chimei-g101ICE_12018182.dts
- imx6dl-dmo-ppc-chimei-g121I1_12018381.dts
- imx6dl-dmo-ppc-HDMI.dts

(it is the same for quad/dual)

when user wants to use customized device tree(oftree), we create a copy of
"user-oftree" with name of "oftree". Then user can apply the LVDS display timing
changes on this file. In the next boot, Barebox will use "oftree" for booting.

"user-oftree" is a copy of our default dts. and only be used for being copied.

Signed-off-by: reyhanehy <reyhane.y84@gmail.com>