swu: symlink and keeping all old images in boot partitin are removed
Previously, when a kernel or oftree updated, a copy of the old one on the
boot partition was created. Therefore, after some updates, there were many old
files on the boot partition, which consumed the disk space.

Also, there was a mechanism which create a symlink to the latest image. Symlink
command returns error(reason not identified).

New idea is implemented:
only one backup is created from the last image before updating. This file is
specified with "-old" postfix to the image filename. And then
the new image is copied to the boot partition.

Signed-off-by: reyhanehy <reyhane.y84@gmail.com>